// lib/models/types.ts
export interface Categoria {
  id_categoria: string;
  nombre: string;
  imagen?: string;
}

// ============================================================
// TABLA PRODUCTO
// ============================================================
export interface Producto {
  id_producto: string;
  nombre: string;
  descripcion: string;
  precio: number;
  id_categoria: string;
  imagen?: string;
  categoria?: Categoria;
  tallas?: string[];
}
// ...
export interface FiltrosProducto {
  categoria?: string;
  precio_min?: number;
  precio_max?: number;
  busqueda?: string;
  orden?: 'precio_asc' | 'precio_desc' | 'nombre' | 'reciente';
}